Module heim_runtime::fs

Structs

File

An open file on the filesystem.

Functions

path_exists

read

Reads the entire contents of a file as raw bytes.

read_dir

Returns a stream of entries in a directory.

read_first_line

read_into

read_lines

read_lines_into

read_link

Reads a symbolic link and returns the path it points to.

read_to_string

Reads the entire contents of a file as a string.